Line example updates, PR comments addressed
