+++ cpp/modules/deck.gl/core/src/lib/layer.cpp
@@ -324,6 +324,12 @@ void Layer::updateState(const Layer::ChangeFlags& changeFlags, const Layer::Prop
 // App can destroy WebGL resources here
 void Layer::finalizeState() {}
 
+void Layer::drawState(wgpu::RenderPassEncoder pass) {
+  for (auto model : this->getModels()) {
+    model->draw(pass);
+  }
+}
+
 // called to populate the info object that is passed to the event handler
 // @return null to cancel event
 // getPickingInfo({info, mode}) {
@@ -358,12 +364,6 @@ void Layer::invalidateAttribute(const std::string& name, const std::string& diff
 //     model.setAttributes(shaderAttributes);
 //   }
 // }
-
-void Layer::_drawState(wgpu::RenderPassEncoder pass) {
-  for (auto model : this->getModels()) {
-    model->draw(pass);
-  }
-}
 
 // LAYER MANAGER API
 // Should only be called by the deck.gl LayerManager class
@@ -441,7 +441,7 @@ void Layer::finalize() {
 // Calculates uniforms
 void Layer::draw(wgpu::RenderPassEncoder pass) {
   // Call subclass lifecycle method
-  this->_drawState(pass);
+  this->drawState(pass);
   // End lifecycle method
 }
