+++ cpp/modules/deck.gl/layers/src/line-layer/line-layer-fragment.glsl.h
@@ -18,6 +18,9 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
+#ifndef DECKGL_LAYERS_LINE_LAYER_FRAGMENT_H
+#define DECKGL_LAYERS_LINE_LAYER_FRAGMENT_H
+
 static const char* fs = R"GLSL(
 #version 450
 
@@ -49,3 +52,5 @@ void main(void) {
 }
 )GLSL";
 */
+
+#endif  // DECKGL_LAYERS_LINE_LAYER_FRAGMENT_H
